* compiling fix for visual studio

+++ otserv/trunk/npc.cpp
@@ -2660,7 +2660,7 @@ int32_t Npc::matchKeywords(NpcResponse* response, std::vector<std::string> wordL
 				//TODO: Should iterate through each word until a number or a new keyword is found.
 				int32_t amount = atoi(lastWordMatchIter->c_str());
 				if(amount > 0){
-					response->setAmount(std::min(amount, 500));
+					response->setAmount(std::min((int32_t)amount, (int32_t)500));
 				}
 				else{
 					response->setAmount(1);
